Mostly done
Mostly done apart from some minor issues:
I wasn't able to get the image looking quite like it does in the mock
The image is not aligned quite right either
The image isn't editable. This would require quite a bit of work to fix.

Otherwise I think this is OK.